backend/requirements.txt
channels
channels-redis
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg
redis